Review Change: extract request and return type of screenshot routes to @models/common/Screenshots
